+++ packages/jobs-pg-boss/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @porulle/jobs-pg-boss
 
+## 0.10.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @porulle/core@0.10.2
+
 ## 0.10.1
 
 ### Patch Changes